->Implemented 2 APIs and Modified 1 API
1) Implemented POST /api/fm/outlets/createOutlet with State, City & Area ID-Based Address Validation, Merchant-Specific Bank Account Validation, Outlet Registration, Address Creation, Bank Details, Operating Days Configuration, and Outlet User Creation.

2) Implemented PUT /api/fm/outlets/updateOutletDetailsByMerchant with Complete Outlet Management Support Including Outlet Information, Contact Details, Address, Bank Details, Location, Operating Days Updates, and Merchant-Specific Bank Account Validation.

3) Modified GET /api/fm/outlets/getOutletDetails to Return Enhanced Outlet Information Including Outlet Email, Alternate Outlet Phone, Bank Details, Complete Address Information, Outlet Timings, Categories, Products, and Favourite Status Based on User Type.

+++ foodandmart/src/main/java/com/jippy/foodandmart/service/IFmMerchantService.java
@@ -18,7 +18,7 @@ public interface IFmMerchantService {
 
     FmBulkUploadResultDTO bulkUpload(MultipartFile file);
     // Get--> merchant + bank
-    FmMerchantWithBankDto getMerchantWithBank(Long merchantId);
+    FmMerchantWithBankDto getMerchantWithBank(Integer merchantId);
 
     // Update--> merchant + bank
     FmMerchantWithBankDto updateMerchantProfile(FmMerchantWithBankDto dto);
